Split out provider tests, new mintr url.
